fix: circuit_breaker record_outbound u32→usize type mismatch (E0308)
estimate_tokens() returns u32 but record_outbound expects usize.
Cast with `as usize` to fix build with dashboard feature enabled.

+++ crates/duduclaw-gateway/src/channel_reply.rs
@@ -485,7 +485,7 @@ async fn build_reply_with_session_inner(
         // Record outbound for circuit breaker echo detection
         let reply_tokens = estimate_tokens(&reply);
         if let Some(ref cb_registry) = ctx.circuit_breakers {
-            cb_registry.record_outbound(session_id, &reply, reply_tokens).await;
+            cb_registry.record_outbound(session_id, &reply, reply_tokens as usize).await;
         }
 
         // Inject defensive prompt if circuit breaker is in HalfOpen (bot loop suspected)
